Split out metadata edit functions
